Fixed bug with refresh token

+++ src/app/shared/services/authservices/authguard.service.ts
@@ -4,33 +4,42 @@ import { UiService } from '../ui.service';
 import { UserService } from '../user.service';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AuthguardService {
+  constructor(
+    public ui: UiService,
+    public userService: UserService,
+    private router: Router,
+    private nav: UiService
+  ) {}
 
-  constructor(public ui: UiService,public userService: UserService, private router: Router, private nav: UiService) { }
-
-  gettoken(){  
+  gettoken() {
     //checks if they have a token and if it is valid, if it is it will refresh the token
-    if(!!localStorage.getItem("UserToken")){
-      this.userService.refreshUser().subscribe( 
-        res => {
-          console.log(res);
-          if(res.success && res.data.token){
-            localStorage.clear();
-            localStorage.setItem('UserToken', res.data.token);
-            console.log(res)
+    if (!!localStorage.getItem('UserToken')) {
+      if (Date.now() >= parseInt(localStorage.getItem('tokenExpiration'))) {
+        this.userService.refreshUser().subscribe(
+          (res) => {
+            console.log(res);
+            if (res.success && res.data.token) {
+              localStorage.clear();
+              localStorage.setItem('UserToken', res.data.token);
+              console.log(res);
+            }
+          },
+          (res) => {
+            if (!res.error.success && res.error.message) {
+              this.ui.showToastMessage('Please log back in', 'danger');
+              localStorage.clear();
+              this.nav.hide();
+              this.router.navigate(['/login']);
+            }
           }
-        },
-        res => {
-          if(!res.error.success && res.error.message) {
-            this.ui.showToastMessage('Please log back in', 'danger');
-            localStorage.clear();
-            this.nav.hide();
-            this.router.navigate(['/login']);
-          }
-        });
+        );
+      } else {
+        console.log('Token NOT expired');
+      }
     }
-    return !!localStorage.getItem("UserToken");  
-    }  
+    return !!localStorage.getItem('UserToken');
+  }
 }
